+++ classes/capitalize_decorator.rb
